+++ src/vm/costs/mod.rs
@@ -132,6 +132,9 @@ impl LimitedCostTracker {
         // used by the miner to "undo" the cost of a transaction when trying to pack a block.
         self.total = total;
     }
+    pub fn get_limit(&self) -> ExecutionCost {
+        self.limit.clone()
+    }
 }
 
 fn add_cost(
