+++ src/vm/functions/mod.rs
@@ -43,7 +43,7 @@ use vm::{eval, Environment, LocalContext};
 use address::AddressHashMode;
 use chainstate::stacks::{StacksAddress, C32_ADDRESS_VERSION_TESTNET_SINGLESIG};
 pub use vm::functions::assets::stx_transfer_consolidated;
-pub use vm::functions::special::{handle_contract_call_special_cases, handle_poison_microblock};
+pub use vm::functions::special::handle_contract_call_special_cases;
 
 define_named_enum!(NativeFunctions {
     Add("+"),
